+++ FDS/3DS/FUtils.h
@@ -11,7 +11,7 @@ void SwapYZ(Quaternion *Q);
 void SwapYZ(Matrix Mat);
 Object *FindObject(short Number);
 Object *FindObject(char *Name);
-void FSeek(long Where);
+void FSeek(int32_t Where);
 DWord FTell();
 void Read(void *Ptr,DWord Size);
 void Compute_UVWrapping_Coordinates(TriMesh *T);
